+++ MelonLoader/Melons/CompatibilityLayers/IPA/IEnhancedPlugin.cs
@@ -1,0 +1,8 @@
+﻿namespace IllusionPlugin
+{
+    public interface IEnhancedPlugin : IPlugin
+    {
+        string[] Filter { get; }
+        void OnLateUpdate();
+    }
+}